Consolidated duped Tabs component
